+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/__tests__/ai-providers-json.spec.ts
@@ -1,5 +1,5 @@
 import defaultAiProviders from 'src/engine/metadata-modules/ai/ai-models/ai-providers.json';
-import { aiProvidersConfigSchema } from 'src/engine/metadata-modules/ai/ai-models/types/ai-providers-config.schema';
+import { aiProvidersConfigSchema } from 'src/engine/metadata-modules/ai/ai-models/types/ai-providers-config.type';
 import { type AiProvidersConfig } from 'src/engine/metadata-modules/ai/ai-models/types/ai-providers-config.type';
 import { buildCompositeModelId } from 'src/engine/metadata-modules/ai/ai-models/utils/composite-model-id.util';
 import { normalizeAiProviders } from 'src/engine/metadata-modules/ai/ai-models/utils/normalize-ai-providers.util';
