+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/types/ai-provider-auth-type.type.ts
@@ -1,1 +1,5 @@
-export type AiProviderAuthType = 'key' | 'credentials' | 'role';
+import { z } from 'zod';
+
+export const aiProviderAuthTypeSchema = z.enum(['key', 'credentials', 'role']);
+
+export type AiProviderAuthType = z.infer<typeof aiProviderAuthTypeSchema>;
